add preset printer, fix somes functions, fix printf negative number !

+++ src/printf/print_nb.c
@@ -6,7 +6,7 @@
 /*   By: aviscogl <aviscogl@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2017/11/24 20:16:41 by alexis       #+#   ##    ##    #+#       */
-/*   Updated: 2017/12/09 16:22:50 by aviscogl    ###    #+. /#+    ###.fr     */
+/*   Updated: 2017/12/16 10:13:51 by aviscogl    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -18,7 +18,7 @@ int		print_nb(int fd, char *p, va_list list)
 {
 	if ((*++p) == 'd' || (*p) == 'i')
 	{
-		ft_putnum_fd(fd, va_arg(list, int64_t), 10);
+		ft_putnum_fd(fd, va_arg(list, int), 10);
 		return (1);
 	}
 	return (0);
